Files refactoring
