+++ src/components/SongItem.tsx
@@ -1,7 +1,7 @@
 import "./SongItem.css"
-import QueueMusic from "./../assets/queue_music.svg"
-import Drag from "./../assets/drag.svg"
-import Remove from "./../assets/close.svg"
+import QueueMusic from "./../assets/icons/queue_music.svg"
+import Drag from "./../assets/icons/drag.svg"
+import Remove from "./../assets/icons/close.svg"
 import usePlaylistStore, { Song } from "../store/playlistStore"
 import { useEffect } from "react"
 import { useSortable } from "@dnd-kit/sortable"
